feat(config): WebSocket / gRPC / SSE per-protocol limits (#641)
Add config-level runtime limits for WebSocket (frame size, deflate),
gRPC (message size), SSE (event size). Pre-wires BuildConfig fields
that USK-643 (Upgrade swap orchestrator) and h2 dispatch will consume
to construct Layers with the configured caps.

* internal/config/protocol_limits.go: new WebSocketLimits / GRPCLimits /
  SSELimits types; resolver helpers (nil/zero -> default); ValidateProtocolLimits
* internal/config/limits.go: add MaxWebSocketFrameSize=16MiB (single
  source of truth; previously private maxFramePayloadSize in ws/frame.go)
* internal/config/config.go: ProxyConfig adds WebSocket/GRPC/SSE pointer
  fields with JSON omitempty (backward compatible)
* internal/layer/ws/frame.go: const moved to config.MaxWebSocketFrameSize
* internal/layer/grpc, grpcweb: add Option type + WithMaxMessageSize,
  change Wrap signature to accept ...Option, replace 5 hardcoded
  config.MaxGRPCMessageSize references with per-channel cap
* internal/connector/stack_builder.go: BuildConfig fields for the four
  new caps; threaded via h2_dispatch into grpc.Wrap call site
* cmd/yorishiro-proxy/main.go: ValidateProtocolLimits hook in init

Project convention: zero = use default (matches BodySpillThreshold);
issue text "reject zero" softened to "reject negative".

Refs: USK-649

+++ internal/layer/ws/deflate.go
@@ -88,7 +88,7 @@ func (ds *deflateState) decompress(payload []byte, maxSize int64) ([]byte, error
 	// Enforce a hard upper bound on the compressed payload size to avoid
 	// pathological allocations and satisfy static analysis that this value
 	// cannot be arbitrarily large.
-	if len(payload) > maxCompressedPayloadSize {
+	if int64(len(payload)) > maxCompressedPayloadSize {
 		return nil, fmt.Errorf("deflate decompress: compressed payload too large: %d > %d", len(payload), maxCompressedPayloadSize)
 	}
